+++ engine/flink/management/src/main/scala/pl/touk/nussknacker/engine/management/FlinkDeploymentManager.scala
@@ -9,7 +9,7 @@ import pl.touk.nussknacker.engine.api.deployment.DeploymentUpdateStrategy.StateR
 import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.deployment.inconsistency.InconsistentStateDetector
 import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
-import pl.touk.nussknacker.engine.api.process.{ProcessIdWithName, ProcessName}
+import pl.touk.nussknacker.engine.api.process.{ProcessIdWithName, ProcessName, VersionId}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.deployment.{
   CustomActionDefinition,
@@ -42,7 +42,9 @@ abstract class FlinkDeploymentManager(
   override def resolve(
       idWithName: ProcessIdWithName,
       statusDetails: List[StatusDetails],
-      lastStateAction: Option[ProcessAction]
+      lastStateAction: Option[ProcessAction],
+      latestVersionId: VersionId,
+      deployedVersionId: Option[VersionId],
   ): Future[ProcessState] = {
     for {
       actionAfterPostprocessOpt <- postprocess(idWithName, statusDetails)
@@ -50,7 +52,11 @@ abstract class FlinkDeploymentManager(
         statusDetails,
         actionAfterPostprocessOpt.orElse(lastStateAction)
       )
-    } yield processStateDefinitionManager.processState(engineStateResolvedWithLastAction)
+    } yield processStateDefinitionManager.processState(
+      engineStateResolvedWithLastAction,
+      latestVersionId,
+      deployedVersionId
+    )
   }
 
   // Flink has a retention for job overviews so we can't rely on this to distinguish between statuses:
@@ -113,7 +119,8 @@ abstract class FlinkDeploymentManager(
         }
       case DMTestScenarioCommand(_, canonicalProcess, scenarioTestData) =>
         testRunner.test(canonicalProcess, scenarioTestData)
-      case command: DMCustomActionCommand => processCustomAction(command)
+      case command: DMCustomActionCommand     => processCustomAction(command)
+      case _: DMPerformSingleExecutionCommand => notImplemented
     }
 
   private def validate(command: DMValidateScenarioCommand): Future[Unit] = {